move extension methods around
